poromech cleanup
