+++ src/coreComponents/physicsSolvers/multiphysics/MultiphasePoromechanics.cpp
@@ -22,7 +22,6 @@
 
 #include "constitutive/fluid/multifluid/MultiFluidBase.hpp"
 #include "constitutive/solid/PorousSolid.hpp"
-#include "mesh/utilities/AverageOverQuadraturePointsKernel.hpp"
 #include "physicsSolvers/fluidFlow/CompositionalMultiphaseBase.hpp"
 #include "physicsSolvers/fluidFlow/FlowSolverBaseFields.hpp"
 #include "physicsSolvers/multiphysics/CompositionalMultiphaseReservoirAndWells.hpp"
@@ -72,8 +71,7 @@ catalogName()
 template< typename FLOW_SOLVER >
 MultiphasePoromechanics< FLOW_SOLVER >::MultiphasePoromechanics( const string & name,
                                                                  Group * const parent )
-  : Base( name, parent ),
-  m_isThermal( 0 )
+  : Base( name, parent )
 {
   this->registerWrapper( viewKeyStruct::stabilizationTypeString(), &m_stabilizationType ).
     setInputFlag( InputFlags::OPTIONAL ).
@@ -92,16 +90,6 @@ MultiphasePoromechanics< FLOW_SOLVER >::MultiphasePoromechanics( const string & 
     setInputFlag( InputFlags::OPTIONAL ).
     setDescription( "Constant multiplier of stabilization strength." );
 
-  this->registerWrapper( viewKeyStruct::isThermalString(), &m_isThermal ).
-    setApplyDefaultValue( 0 ).
-    setInputFlag( InputFlags::OPTIONAL ).
-    setDescription( "Flag indicating whether the problem is thermal or not. Set isThermal=\"1\" to enable the thermal coupling" );
-
-  this->registerWrapper( viewKeyStruct::performStressInitializationString(), &m_performStressInitialization ).
-    setApplyDefaultValue( false ).
-    setInputFlag( InputFlags::FALSE ).
-    setDescription( "Flag to indicate that the solver is going to perform stress initialization" );
-
   LinearSolverParameters & linearSolverParameters = this->m_linearSolverParameters.get();
   linearSolverParameters.mgr.strategy = LinearSolverParameters::MGR::StrategyType::multiphasePoromechanics;
   linearSolverParameters.mgr.separateComponents = true;
@@ -114,7 +102,7 @@ void MultiphasePoromechanics< FLOW_SOLVER >::postProcessInput()
 {
   Base::postProcessInput();
 
-  GEOS_ERROR_IF( flowSolver()->catalogName() == "CompositionalMultiphaseReservoir" &&
+  GEOS_ERROR_IF( this->flowSolver()->catalogName() == "CompositionalMultiphaseReservoir" &&
                  this->getNonlinearSolverParameters().couplingType() != NonlinearSolverParameters::CouplingType::Sequential,
                  GEOS_FMT( "{}: {} solver is only designed to work for {} = {}",
                            this->getDataContext(), catalogName(), NonlinearSolverParameters::viewKeysStruct::couplingTypeString(),
@@ -125,46 +113,26 @@ void MultiphasePoromechanics< FLOW_SOLVER >::postProcessInput()
 template< typename FLOW_SOLVER >
 void MultiphasePoromechanics< FLOW_SOLVER >::registerDataOnMesh( Group & meshBodies )
 {
-  SolverBase::registerDataOnMesh( meshBodies );
+  Base::registerDataOnMesh( meshBodies );
 
-  if( this->getNonlinearSolverParameters().m_couplingType == NonlinearSolverParameters::CouplingType::Sequential )
+  if( m_stabilizationType == StabilizationType::Global ||
+      m_stabilizationType == StabilizationType::Local )
   {
-    // to let the solid mechanics solver that there is a pressure and temperature RHS in the mechanics solve
-    solidMechanicsSolver()->enableFixedStressPoromechanicsUpdate();
-    // to let the flow solver that saving pressure_k and temperature_k is necessary (for the fixed-stress porosity terms)
-    flowSolver()->enableFixedStressPoromechanicsUpdate();
-  }
-
-  this->template forDiscretizationOnMeshTargets( meshBodies, [&] ( string const &,
-                                                                   MeshLevel & mesh,
-                                                                   arrayView1d< string const > const & regionNames )
-  {
-    ElementRegionManager & elemManager = mesh.getElemManager();
-
-    elemManager.forElementSubRegions< ElementSubRegionBase >( regionNames,
-                                                              [&]( localIndex const,
-                                                                   ElementSubRegionBase & subRegion )
+    this->template forDiscretizationOnMeshTargets( meshBodies, [&] ( string const &,
+                                                                     MeshLevel & mesh,
+                                                                     arrayView1d< string const > const & regionNames )
     {
-      subRegion.registerWrapper< string >( viewKeyStruct::porousMaterialNamesString() ).
-        setPlotLevel( PlotLevel::NOPLOT ).
-        setRestartFlags( RestartFlags::NO_WRITE ).
-        setSizedFromParent( 0 );
+      ElementRegionManager & elemManager = mesh.getElemManager();
 
-      if( m_stabilizationType == StabilizationType::Global ||
-          m_stabilizationType == StabilizationType::Local )
+      elemManager.forElementSubRegions< ElementSubRegionBase >( regionNames,
+                                                                [&]( localIndex const,
+                                                                     ElementSubRegionBase & subRegion )
       {
         subRegion.registerField< fields::flow::macroElementIndex >( this->getName() );
         subRegion.registerField< fields::flow::elementStabConstant >( this->getName() );
-      }
-
-      if( this->getNonlinearSolverParameters().m_couplingType == NonlinearSolverParameters::CouplingType::Sequential )
-      {
-        // register the bulk density for use in the solid mechanics solver
-        // ideally we would resize it here as well, but the solid model name is not available yet (see below)
-        subRegion.registerField< fields::poromechanics::bulkDensity >( this->getName() );
-      }
+      } );
     } );
-  } );
+  }
 }
 
 template< typename FLOW_SOLVER >
@@ -175,20 +143,6 @@ void MultiphasePoromechanics< FLOW_SOLVER >::setupCoupling( DomainPartition cons
                           CompositionalMultiphaseBase::viewKeyStruct::elemDofFieldString(),
                           DofManager::Connector::Elem );
 }
-
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::setupDofs( DomainPartition const & domain,
-                                                        DofManager & dofManager ) const
-{
-  // note that the order of operations matters a lot here (for instance for the MGR labels)
-  // we must set up dofs for solid mechanics first, and then for flow
-  // that's the reason why this function is here and not in CoupledSolvers.hpp
-  solidMechanicsSolver()->setupDofs( domain, dofManager );
-  flowSolver()->setupDofs( domain, dofManager );
-
-  setupCoupling( domain, dofManager );
-}
-
 
 template< typename FLOW_SOLVER >
 void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_UNUSED_PARAM( time ),
@@ -215,7 +169,7 @@ void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_U
 
     string const flowDofKey = dofManager.getKey( CompositionalMultiphaseBase::viewKeyStruct::elemDofFieldString() );
 
-    if( m_isThermal )
+    if( this->m_isThermal )
     {
       poromechanicsMaxForce =
         assemblyLaunch< constitutive::PorousSolid< ElasticIsotropic >, // TODO: change once there is a cmake solution
@@ -227,9 +181,9 @@ void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_U
                                                                                                     localRhs,
                                                                                                     dt,
                                                                                                     flowDofKey,
-                                                                                                    flowSolver()->numFluidComponents(),
-                                                                                                    flowSolver()->numFluidPhases(),
-                                                                                                    flowSolver()->useTotalMassEquation(),
+                                                                                                    this->flowSolver()->numFluidComponents(),
+                                                                                                    this->flowSolver()->numFluidPhases(),
+                                                                                                    this->flowSolver()->useTotalMassEquation(),
                                                                                                     FlowSolverBase::viewKeyStruct::fluidNamesString() );
     }
     else
@@ -244,18 +198,18 @@ void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_U
                                                                                       localRhs,
                                                                                       dt,
                                                                                       flowDofKey,
-                                                                                      flowSolver()->numFluidComponents(),
-                                                                                      flowSolver()->numFluidPhases(),
-                                                                                      flowSolver()->useTotalMassEquation(),
+                                                                                      this->flowSolver()->numFluidComponents(),
+                                                                                      this->flowSolver()->numFluidPhases(),
+                                                                                      this->flowSolver()->useTotalMassEquation(),
                                                                                       FlowSolverBase::viewKeyStruct::fluidNamesString() );
     }
   } );
 
   // step 2: apply mechanics solver on its target regions not included in the poromechanics solver target regions
 
-  solidMechanicsSolver()->forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                                        MeshLevel & mesh,
-                                                                                        arrayView1d< string const > const & regionNames )
+  this->solidMechanicsSolver()->forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
+                                                                                              MeshLevel & mesh,
+                                                                                              arrayView1d< string const > const & regionNames )
   {
 
     // collect the target region of the mechanics solver not included in the poromechanics target regions
@@ -289,7 +243,7 @@ void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_U
   } );
 
 
-  solidMechanicsSolver()->getMaxForce() = LvArray::math::max( mechanicsMaxForce, poromechanicsMaxForce );
+  this->solidMechanicsSolver()->getMaxForce() = LvArray::math::max( mechanicsMaxForce, poromechanicsMaxForce );
 
   // step 3: compute the fluxes (face-based contributions)
 
@@ -297,19 +251,19 @@ void MultiphasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const GEOS_U
       m_stabilizationType == StabilizationType::Local )
   {
     updateStabilizationParameters( domain );
-    flowSolver()->assembleStabilizedFluxTerms( dt,
-                                               domain,
-                                               dofManager,
-                                               localMatrix,
-                                               localRhs );
+    this->flowSolver()->assembleStabilizedFluxTerms( dt,
+                                                     domain,
+                                                     dofManager,
+                                                     localMatrix,
+                                                     localRhs );
   }
   else
   {
-    flowSolver()->assembleFluxTerms( dt,
-                                     domain,
-                                     dofManager,
-                                     localMatrix,
-                                     localRhs );
+    this->flowSolver()->assembleFluxTerms( dt,
+                                           domain,
+                                           dofManager,
+                                           localMatrix,
+                                           localRhs );
   }
 }
 
@@ -326,11 +280,11 @@ void MultiphasePoromechanics< FLOW_SOLVER >::updateState( DomainPartition & doma
                                                               [&]( localIndex const,
                                                                    CellElementSubRegion & subRegion )
     {
-      real64 const deltaPhaseVolFrac = flowSolver()->updateFluidState( subRegion );
+      real64 const deltaPhaseVolFrac = this->flowSolver()->updateFluidState( subRegion );
       maxDeltaPhaseVolFrac = LvArray::math::max( maxDeltaPhaseVolFrac, deltaPhaseVolFrac );
-      if( m_isThermal )
+      if( this->m_isThermal )
       {
-        flowSolver()->updateSolidInternalEnergyModel( subRegion );
+        this->flowSolver()->updateSolidInternalEnergyModel( subRegion );
       }
     } );
   } );
@@ -342,14 +296,14 @@ void MultiphasePoromechanics< FLOW_SOLVER >::updateState( DomainPartition & doma
 template< typename FLOW_SOLVER >
 void MultiphasePoromechanics< FLOW_SOLVER >::initializePostInitialConditionsPreSubGroups()
 {
-  SolverBase::initializePostInitialConditionsPreSubGroups();
+  Base::initializePostInitialConditionsPreSubGroups();
 
   arrayView1d< string const > const & poromechanicsTargetRegionNames =
     this->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
   arrayView1d< string const > const & solidMechanicsTargetRegionNames =
-    solidMechanicsSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
+    this->solidMechanicsSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
   arrayView1d< string const > const & flowTargetRegionNames =
-    flowSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
+    this->flowSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
   for( integer i = 0; i < poromechanicsTargetRegionNames.size(); ++i )
   {
     GEOS_THROW_IF( std::find( solidMechanicsTargetRegionNames.begin(), solidMechanicsTargetRegionNames.end(),
@@ -357,22 +311,22 @@ void MultiphasePoromechanics< FLOW_SOLVER >::initializePostInitialConditionsPreS
                    == solidMechanicsTargetRegionNames.end(),
                    GEOS_FMT( "{} {}: region {} must be a target region of {}",
                              getCatalogName(), this->getDataContext(), poromechanicsTargetRegionNames[i],
-                             solidMechanicsSolver()->getDataContext() ),
+                             this->solidMechanicsSolver()->getDataContext() ),
                    InputError );
     GEOS_THROW_IF( std::find( flowTargetRegionNames.begin(), flowTargetRegionNames.end(), poromechanicsTargetRegionNames[i] )
                    == flowTargetRegionNames.end(),
                    GEOS_FMT( "{} {}: region `{}` must be a target region of `{}`",
-                             getCatalogName(), this->getDataContext(), poromechanicsTargetRegionNames[i], flowSolver()->getDataContext() ),
+                             getCatalogName(), this->getDataContext(), poromechanicsTargetRegionNames[i], this->flowSolver()->getDataContext() ),
                    InputError );
   }
 
-  integer & isFlowThermal = flowSolver()->isThermal();
-  GEOS_WARNING_IF( m_isThermal && !isFlowThermal,
+  integer & isFlowThermal = this->flowSolver()->isThermal();
+  GEOS_WARNING_IF( this->m_isThermal && !isFlowThermal,
                    GEOS_FMT( "{} {}: The attribute `{}` of the flow solver `{}` is set to 1 since the poromechanics solver is thermal",
-                             getCatalogName(), this->getName(), FlowSolverBase::viewKeyStruct::isThermalString(), flowSolver()->getName() ) );
-  isFlowThermal = m_isThermal;
+                             getCatalogName(), this->getName(), FlowSolverBase::viewKeyStruct::isThermalString(), this->flowSolver()->getName() ) );
+  isFlowThermal = this->m_isThermal;
 
-  if( m_isThermal )
+  if( this->m_isThermal )
   {
     this->m_linearSolverParameters.get().mgr.strategy = LinearSolverParameters::MGR::StrategyType::thermalMultiphasePoromechanics;
   }
@@ -381,46 +335,12 @@ void MultiphasePoromechanics< FLOW_SOLVER >::initializePostInitialConditionsPreS
 template< typename FLOW_SOLVER >
 void MultiphasePoromechanics< FLOW_SOLVER >::initializePreSubGroups()
 {
-  SolverBase::initializePreSubGroups();
+  Base::initializePreSubGroups();
 
   GEOS_THROW_IF( m_stabilizationType == StabilizationType::Local,
                  this->getWrapperDataContext( viewKeyStruct::stabilizationTypeString() ) <<
                  ": Local stabilization has been disabled temporarily",
                  InputError );
-
-  DomainPartition & domain = this->template getGroupByPath< DomainPartition >( "/Problem/domain" );
-
-  this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                               MeshLevel & mesh,
-                                                                               arrayView1d< string const > const & regionNames )
-  {
-    ElementRegionManager & elementRegionManager = mesh.getElemManager();
-    elementRegionManager.forElementSubRegions< ElementSubRegionBase >( regionNames,
-                                                                       [&]( localIndex const,
-                                                                            ElementSubRegionBase & subRegion )
-    {
-      string & porousName = subRegion.getReference< string >( viewKeyStruct::porousMaterialNamesString() );
-      porousName = this->template getConstitutiveName< CoupledSolidBase >( subRegion );
-      GEOS_ERROR_IF( porousName.empty(), GEOS_FMT( "{}: Solid model not found on subregion {}",
-                                                   this->getDataContext(), subRegion.getName() ) );
-
-      if( subRegion.hasField< fields::poromechanics::bulkDensity >() )
-      {
-        // get the solid model to know the number of quadrature points and resize the bulk density
-        CoupledSolidBase const & solid = this->template getConstitutiveModel< CoupledSolidBase >( subRegion, porousName );
-        subRegion.getField< fields::poromechanics::bulkDensity >().resizeDimension< 1 >( solid.getDensity().size( 1 ) );
-      }
-    } );
-  } );
-}
-
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::implicitStepSetup( real64 const & time_n,
-                                                                real64 const & dt,
-                                                                DomainPartition & domain )
-{
-  flowSolver()->keepFlowVariablesConstantDuringInitStep( m_performStressInitialization );
-  Base::implicitStepSetup( time_n, dt, domain );
 }
 
 template< typename FLOW_SOLVER >
@@ -488,57 +408,6 @@ void MultiphasePoromechanics< FLOW_SOLVER >::updateStabilizationParameters( Doma
 }
 
 template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::mapSolutionBetweenSolvers( DomainPartition & domain, integer const solverType )
-{
-  GEOS_MARK_FUNCTION;
-
-  /// After the flow solver
-  if( solverType == static_cast< integer >( Base::SolverType::Flow ) )
-  {
-    // save pressure and temperature at the end of this iteration
-    flowSolver()->saveIterationState( domain );
-
-    this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                                MeshLevel & mesh,
-                                                                                arrayView1d< string const > const & regionNames )
-    {
-
-      mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                            auto & subRegion )
-      {
-        // update the bulk density
-        // TODO: ideally, we would not recompute the bulk density, but a more general "rhs" containing the body force and the
-        // pressure/temperature terms
-        updateBulkDensity( subRegion );
-      } );
-    } );
-  }
-
-  /// After the solid mechanics solver
-  if( solverType == static_cast< integer >( Base::SolverType::SolidMechanics ) )
-  {
-    // compute the average of the mean total stress increment over quadrature points
-    averageMeanTotalStressIncrement( domain );
-
-    this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                                MeshLevel & mesh,
-                                                                                arrayView1d< string const > const & regionNames )
-    {
-
-      mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                            auto & subRegion )
-      {
-        // update the porosity after a change in displacement (after mechanics solve)
-        // or a change in pressure/temperature (after a flow solve)
-        flowSolver()->updatePorosityAndPermeability( subRegion );
-      } );
-    } );
-  }
-  // call base method (needed to perform nonlinear acceleration)
-  Base::mapSolutionBetweenSolvers( domain, solverType );
-}
-
-template< typename FLOW_SOLVER >
 void MultiphasePoromechanics< FLOW_SOLVER >::updateBulkDensity( ElementSubRegionBase & subRegion )
 {
   // get the fluid model (to access fluid density)
@@ -552,52 +421,10 @@ void MultiphasePoromechanics< FLOW_SOLVER >::updateBulkDensity( ElementSubRegion
   // update the bulk density
   poromechanicsKernels::
     MultiphaseBulkDensityKernelFactory::
-    createAndLaunch< parallelDevicePolicy<> >( flowSolver()->numFluidPhases(),
+    createAndLaunch< parallelDevicePolicy<> >( this->flowSolver()->numFluidPhases(),
                                                fluid,
                                                solid,
                                                subRegion );
-}
-
-template< typename FLOW_SOLVER >
-void MultiphasePoromechanics< FLOW_SOLVER >::averageMeanTotalStressIncrement( DomainPartition & domain )
-{
-  this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                              MeshLevel & mesh,
-                                                                              arrayView1d< string const > const & regionNames )
-  {
-    mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                          auto & subRegion )
-    {
-      // get the solid model (to access stress increment)
-      string const solidName = subRegion.template getReference< string >( viewKeyStruct::porousMaterialNamesString() );
-      CoupledSolidBase & solid = this->template getConstitutiveModel< CoupledSolidBase >( subRegion, solidName );
-
-      arrayView2d< real64 const > const meanTotalStressIncrement_k = solid.getMeanTotalStressIncrement_k();
-      arrayView1d< real64 > const averageMeanTotalStressIncrement_k = solid.getAverageMeanTotalStressIncrement_k();
-
-      finiteElement::FiniteElementBase & subRegionFE =
-        subRegion.template getReference< finiteElement::FiniteElementBase >( solidMechanicsSolver()->getDiscretizationName() );
-
-      // determine the finite element type
-      finiteElement::FiniteElementDispatchHandler< BASE_FE_TYPES >::
-      dispatch3D( subRegionFE, [&] ( auto const finiteElement )
-      {
-        using FE_TYPE = decltype( finiteElement );
-
-        // call the factory and launch the kernel
-        AverageOverQuadraturePoints1DKernelFactory::
-          createAndLaunch< CellElementSubRegion,
-                           FE_TYPE,
-                           parallelDevicePolicy<> >( mesh.getNodeManager(),
-                                                     mesh.getEdgeManager(),
-                                                     mesh.getFaceManager(),
-                                                     subRegion,
-                                                     finiteElement,
-                                                     meanTotalStressIncrement_k,
-                                                     averageMeanTotalStressIncrement_k );
-      } );
-    } );
-  } );
 }
 
 template class MultiphasePoromechanics< CompositionalMultiphaseBase >;
