+++ src/coreComponents/physicsSolvers/multiphysics/MultiphasePoromechanics.hpp
@@ -19,10 +19,8 @@
 #ifndef GEOS_PHYSICSSOLVERS_MULTIPHYSICS_MULTIPHASEPOROMECHANICS_HPP_
 #define GEOS_PHYSICSSOLVERS_MULTIPHYSICS_MULTIPHASEPOROMECHANICS_HPP_
 
-#include "physicsSolvers/multiphysics/CoupledSolver.hpp"
-#include "constitutive/solid/CoupledSolidBase.hpp"
-#include "physicsSolvers/fluidFlow/CompositionalMultiphaseBase.hpp"
 #include "physicsSolvers/multiphysics/PoromechanicsSolver.hpp"
+#include "physicsSolvers/fluidFlow/CompositionalMultiphaseBase.hpp"
 
 
 namespace geos
@@ -77,24 +75,6 @@ public:
   string getCatalogName() const override { return catalogName(); }
 
   /**
-   * @brief accessor for the pointer to the solid mechanics solver
-   * @return a pointer to the solid mechanics solver
-   */
-  SolidMechanicsLagrangianFEM * solidMechanicsSolver() const
-  {
-    return std::get< toUnderlying( Base::SolverType::SolidMechanics ) >( m_solvers );
-  }
-
-  /**
-   * @brief accessor for the pointer to the flow solver
-   * @return a pointer to the flow solver
-   */
-  FLOW_SOLVER * flowSolver() const
-  {
-    return std::get< toUnderlying( Base::SolverType::Flow ) >( m_solvers );
-  }
-
-  /**
    * @defgroup Solver Interface Functions
    *
    * These functions provide the primary interface that is required for derived classes
@@ -107,13 +87,6 @@ public:
 
   virtual void setupCoupling( DomainPartition const & domain,
                               DofManager & dofManager ) const override;
-
-  virtual void setupDofs( DomainPartition const & domain,
-                          DofManager & dofManager ) const override;
-
-  virtual void implicitStepSetup( real64 const & time_n,
-                                  real64 const & dt,
-                                  DomainPartition & domain ) override;
 
   virtual void assembleSystem( real64 const time,
                                real64 const dt,
@@ -132,15 +105,6 @@ public:
    */
   void updateStabilizationParameters( DomainPartition & domain ) const;
 
-  /*
-   * @brief Utility function to set the stress initialization flag
-   * @param[in] performStressInitialization true if the solver has to initialize stress, false otherwise
-   */
-  void setStressInitialization( integer const performStressInitialization )
-  { m_performStressInitialization = performStressInitialization; }
-
-  virtual void mapSolutionBetweenSolvers( DomainPartition & domain, integer const solverType ) override final;
-
 protected:
 
   virtual void initializePostInitialConditionsPreSubGroups() override;
@@ -149,9 +113,6 @@ protected:
 
   struct viewKeyStruct : Base::viewKeyStruct
   {
-    /// Names of the porous materials
-    constexpr static char const * porousMaterialNamesString() { return "porousMaterialNames"; }
-
     /// Type of stabilization used in the simulation
     constexpr static char const * stabilizationTypeString() { return "stabilizationType"; }
 
@@ -160,13 +121,6 @@ protected:
 
     /// Multiplier on stabilization
     constexpr static char const * stabilizationMultiplierString() { return "stabilizationMultiplier"; }
-
-    /// Flag to determine whether or not this is aa thermal simulation
-    constexpr static char const * isThermalString() { return "isThermal"; }
-
-    /// Flag to indicate that the solver is going to perform stress initialization
-    constexpr static char const * performStressInitializationString() { return "performStressInitialization"; }
-
   };
 
 private:
@@ -175,13 +129,7 @@ private:
    * @brief Helper function to recompute the bulk density
    * @param[in] subRegion the element subRegion
    */
-  void updateBulkDensity( ElementSubRegionBase & subRegion );
-
-  /**
-   * @brief Helper function to average the mean total stress increment over quadrature points
-   * @param[in] domain the domain partition
-   */
-  void averageMeanTotalStressIncrement( DomainPartition & domain );
+  virtual void updateBulkDensity( ElementSubRegionBase & subRegion ) override;
 
   template< typename CONSTITUTIVE_BASE,
             typename KERNEL_WRAPPER,
@@ -203,12 +151,6 @@ private:
 
   /// Multiplier on stabilization constant
   real64 m_stabilizationMultiplier;
-
-  /// flag to determine whether or not this is a thermal simulation
-  integer m_isThermal;
-
-  /// Flag to indicate that the solver is going to perform stress initialization
-  integer m_performStressInitialization;
 
 };
 
@@ -247,7 +189,7 @@ real64 MultiphasePoromechanics< FLOW_SOLVER >::assemblyLaunch( MeshLevel & mesh,
                                          CONSTITUTIVE_BASE,
                                          CellElementSubRegion >( mesh,
                                                                  regionNames,
-                                                                 solidMechanicsSolver()->getDiscretizationName(),
+                                                                 this->solidMechanicsSolver()->getDiscretizationName(),
                                                                  materialNamesString,
                                                                  kernelWrapper );
 }
