+++ src/coreComponents/physicsSolvers/multiphysics/SinglePhasePoromechanics.cpp
@@ -24,7 +24,6 @@
 #include "constitutive/fluid/singlefluid/SingleFluidBase.hpp"
 #include "linearAlgebra/solvers/BlockPreconditioner.hpp"
 #include "linearAlgebra/solvers/SeparateComponentPreconditioner.hpp"
-#include "mesh/utilities/AverageOverQuadraturePointsKernel.hpp"
 #include "physicsSolvers/fluidFlow/SinglePhaseBase.hpp"
 #include "physicsSolvers/multiphysics/SinglePhaseReservoirAndWells.hpp"
 #include "physicsSolvers/multiphysics/poromechanicsKernels/SinglePhasePoromechanics.hpp"
@@ -73,19 +72,8 @@ catalogName()
 template< typename FLOW_SOLVER >
 SinglePhasePoromechanics< FLOW_SOLVER >::SinglePhasePoromechanics( const string & name,
                                                                    Group * const parent )
-  : Base( name, parent ),
-  m_isThermal( 0 )
+  : Base( name, parent )
 {
-  this->registerWrapper( viewKeyStruct::isThermalString(), &m_isThermal ).
-    setApplyDefaultValue( 0 ).
-    setInputFlag( InputFlags::OPTIONAL ).
-    setDescription( "Flag indicating whether the problem is thermal or not. Set isThermal=\"1\" to enable the thermal coupling" );
-
-  this->registerWrapper( viewKeyStruct::performStressInitializationString(), &m_performStressInitialization ).
-    setApplyDefaultValue( false ).
-    setInputFlag( InputFlags::FALSE ).
-    setDescription( "Flag to indicate that the solver is going to perform stress initialization" );
-
   LinearSolverParameters & linearSolverParameters = this->m_linearSolverParameters.get();
   linearSolverParameters.mgr.strategy = LinearSolverParameters::MGR::StrategyType::singlePhasePoromechanics;
   linearSolverParameters.mgr.separateComponents = true;
@@ -98,52 +86,12 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::postProcessInput()
 {
   Base::postProcessInput();
 
-  GEOS_ERROR_IF( flowSolver()->catalogName() == "CompositionalMultiphaseReservoir" &&
+  GEOS_ERROR_IF( this->flowSolver()->catalogName() == "SinglePhaseReservoir" &&
                  this->getNonlinearSolverParameters().couplingType() != NonlinearSolverParameters::CouplingType::Sequential,
                  GEOS_FMT( "{}: {} solver is only designed to work for {} = {}",
                            this->getName(), catalogName(), NonlinearSolverParameters::viewKeysStruct::couplingTypeString(),
                            EnumStrings< NonlinearSolverParameters::CouplingType >::toString( NonlinearSolverParameters::CouplingType::Sequential )
                            ));
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::registerDataOnMesh( Group & meshBodies )
-{
-  SolverBase::registerDataOnMesh( meshBodies );
-
-  if( this->getNonlinearSolverParameters().m_couplingType == NonlinearSolverParameters::CouplingType::Sequential )
-  {
-    // to let the solid mechanics solver that there is a pressure and temperature RHS in the mechanics solve
-    solidMechanicsSolver()->enableFixedStressPoromechanicsUpdate();
-    // to let the flow solver that saving pressure_k and temperature_k is necessary (for the fixed-stress porosity terms)
-    flowSolver()->enableFixedStressPoromechanicsUpdate();
-  }
-
-  this->template forDiscretizationOnMeshTargets( meshBodies, [&] ( string const &,
-                                                                   MeshLevel & mesh,
-                                                                   arrayView1d< string const > const & regionNames )
-  {
-
-    ElementRegionManager & elemManager = mesh.getElemManager();
-
-    elemManager.forElementSubRegions< ElementSubRegionBase >( regionNames,
-                                                              [&]( localIndex const,
-                                                                   ElementSubRegionBase & subRegion )
-    {
-      subRegion.registerWrapper< string >( viewKeyStruct::porousMaterialNamesString() ).
-        setPlotLevel( PlotLevel::NOPLOT ).
-        setRestartFlags( RestartFlags::NO_WRITE ).
-        setSizedFromParent( 0 );
-
-      if( this->getNonlinearSolverParameters().m_couplingType == NonlinearSolverParameters::CouplingType::Sequential )
-      {
-        // register the bulk density for use in the solid mechanics solver
-        // ideally we would resize it here as well, but the solid model name is not available yet (see below)
-        subRegion.registerField< fields::poromechanics::bulkDensity >( this->getName() );
-      }
-
-    } );
-  } );
 }
 
 template< typename FLOW_SOLVER >
@@ -153,63 +101,6 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::setupCoupling( DomainPartition con
   dofManager.addCoupling( solidMechanics::totalDisplacement::key(),
                           SinglePhaseBase::viewKeyStruct::elemDofFieldString(),
                           DofManager::Connector::Elem );
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::setupDofs( DomainPartition const & domain,
-                                                         DofManager & dofManager ) const
-{
-  // note that the order of operations matters a lot here (for instance for the MGR labels)
-  // we must set up dofs for solid mechanics first, and then for flow
-  // that's the reason why this function is here and not in CoupledSolvers.hpp
-  solidMechanicsSolver()->setupDofs( domain, dofManager );
-  flowSolver()->setupDofs( domain, dofManager );
-
-  setupCoupling( domain, dofManager );
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::initializePreSubGroups()
-{
-  SolverBase::initializePreSubGroups();
-
-  DomainPartition & domain = this->template getGroupByPath< DomainPartition >( "/Problem/domain" );
-
-  this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                               MeshLevel & mesh,
-                                                                               arrayView1d< string const > const & regionNames )
-  {
-    ElementRegionManager & elementRegionManager = mesh.getElemManager();
-    elementRegionManager.forElementSubRegions< ElementSubRegionBase >( regionNames,
-                                                                       [&]( localIndex const,
-                                                                            ElementSubRegionBase & subRegion )
-    {
-      string & porousName = subRegion.getReference< string >( viewKeyStruct::porousMaterialNamesString() );
-      porousName = this->template getConstitutiveName< CoupledSolidBase >( subRegion );
-      GEOS_THROW_IF( porousName.empty(),
-                     GEOS_FMT( "{} {} : Solid model not found on subregion {}",
-                               getCatalogName(), this->getDataContext().toString(), subRegion.getName() ),
-                     InputError );
-
-      if( subRegion.hasField< fields::poromechanics::bulkDensity >() )
-      {
-        // get the solid model to know the number of quadrature points and resize the bulk density
-        CoupledSolidBase const & solid = this->template getConstitutiveModel< CoupledSolidBase >( subRegion, porousName );
-        subRegion.getField< fields::poromechanics::bulkDensity >().resizeDimension< 1 >( solid.getDensity().size( 1 ) );
-      }
-
-    } );
-  } );
-
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::implicitStepSetup( real64 const & time_n,
-                                                                 real64 const & dt,
-                                                                 DomainPartition & domain )
-{
-  flowSolver()->keepFlowVariablesConstantDuringInitStep( m_performStressInitialization );
-  Base::implicitStepSetup( time_n, dt, domain );
 }
 
 template< typename FLOW_SOLVER >
@@ -237,35 +128,35 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::setupSystem( DomainPartition & dom
 template< typename FLOW_SOLVER >
 void SinglePhasePoromechanics< FLOW_SOLVER >::initializePostInitialConditionsPreSubGroups()
 {
-  SolverBase::initializePostInitialConditionsPreSubGroups();
+  Base::initializePostInitialConditionsPreSubGroups();
 
   arrayView1d< string const > const & poromechanicsTargetRegionNames =
     this->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
   arrayView1d< string const > const & flowTargetRegionNames =
-    flowSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
+    this->flowSolver()->template getReference< array1d< string > >( SolverBase::viewKeyStruct::targetRegionsString() );
   for( integer i = 0; i < poromechanicsTargetRegionNames.size(); ++i )
   {
     GEOS_THROW_IF( std::find( flowTargetRegionNames.begin(), flowTargetRegionNames.end(), poromechanicsTargetRegionNames[i] )
                    == flowTargetRegionNames.end(),
                    GEOS_FMT( "{} {}: region `{}` must be a target region of `{}`",
-                             getCatalogName(), this->getDataContext(), poromechanicsTargetRegionNames[i], flowSolver()->getDataContext() ),
+                             getCatalogName(), this->getDataContext(), poromechanicsTargetRegionNames[i], this->flowSolver()->getDataContext() ),
                    InputError );
   }
 
-  integer & isFlowThermal = flowSolver()->isThermal();
-  GEOS_LOG_RANK_0_IF( m_isThermal && !isFlowThermal,
+  integer & isFlowThermal = this->flowSolver()->isThermal();
+  GEOS_LOG_RANK_0_IF( this->m_isThermal && !isFlowThermal,
                       GEOS_FMT( "{} {}: The attribute `{}` of the flow solver `{}` is set to 1 since the poromechanics solver is thermal",
                                 getCatalogName(), this->getName(),
-                                FlowSolverBase::viewKeyStruct::isThermalString(), flowSolver()->getDataContext() ) );
-  isFlowThermal = m_isThermal;
+                                FlowSolverBase::viewKeyStruct::isThermalString(), this->flowSolver()->getDataContext() ) );
+  isFlowThermal = this->m_isThermal;
 
-  if( m_isThermal )
+  if( this->m_isThermal )
   {
     this->m_linearSolverParameters.get().mgr.strategy = LinearSolverParameters::MGR::StrategyType::thermalSinglePhasePoromechanics;
   }
   else
   {
-    if( flowSolver()->getLinearSolverParameters().mgr.strategy == LinearSolverParameters::MGR::StrategyType::singlePhaseHybridFVM )
+    if( this->flowSolver()->getLinearSolverParameters().mgr.strategy == LinearSolverParameters::MGR::StrategyType::singlePhaseHybridFVM )
     {
       this->m_linearSolverParameters.get().mgr.strategy = LinearSolverParameters::MGR::StrategyType::hybridSinglePhasePoromechanics;
     }
@@ -291,11 +182,11 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::assembleSystem( real64 const time_
                              localRhs );
 
   // Step 3: compute the fluxes (face-based contributions)
-  flowSolver()->assembleFluxTerms( dt,
-                                   domain,
-                                   dofManager,
-                                   localMatrix,
-                                   localRhs );
+  this->flowSolver()->assembleFluxTerms( dt,
+                                         domain,
+                                         dofManager,
+                                         localMatrix,
+                                         localRhs );
 
 }
 
@@ -325,7 +216,7 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::assembleElementBasedTerms( real64 
 
     string const flowDofKey = dofManager.getKey( SinglePhaseBase::viewKeyStruct::elemDofFieldString() );
 
-    if( m_isThermal )
+    if( this->m_isThermal )
     {
       poromechanicsMaxForce =
         assemblyLaunch< constitutive::PorousSolid< ElasticIsotropic >, // TODO: change once there is a cmake solution
@@ -357,9 +248,9 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::assembleElementBasedTerms( real64 
 
   // step 2: apply mechanics solver on its target regions not included in the poromechanics solver target regions
 
-  solidMechanicsSolver()->forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
-                                                                                        MeshLevel & mesh,
-                                                                                        arrayView1d< string const > const & regionNames )
+  this->solidMechanicsSolver()->forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&] ( string const &,
+                                                                                              MeshLevel & mesh,
+                                                                                              arrayView1d< string const > const & regionNames )
   {
     // collect the target region of the mechanics solver not included in the poromechanics target regions
     array1d< string > filteredRegionNames;
@@ -390,7 +281,7 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::assembleElementBasedTerms( real64 
                                                                                 dt );
   } );
 
-  solidMechanicsSolver()->getMaxForce() = LvArray::math::max( mechanicsMaxForce, poromechanicsMaxForce );
+  this->solidMechanicsSolver()->getMaxForce() = LvArray::math::max( mechanicsMaxForce, poromechanicsMaxForce );
 }
 
 template< typename FLOW_SOLVER >
@@ -402,12 +293,12 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::createPreconditioner()
                                                                            SchurComplementOption::RowsumDiagonalProbing,
                                                                            BlockScalingOption::FrobeniusNorm );
 
-    auto mechPrecond = LAInterface::createPreconditioner( solidMechanicsSolver()->getLinearSolverParameters() );
+    auto mechPrecond = LAInterface::createPreconditioner( this->solidMechanicsSolver()->getLinearSolverParameters() );
     precond->setupBlock( 0,
                          { { solidMechanics::totalDisplacement::key(), { 3, true } } },
                          std::make_unique< SeparateComponentPreconditioner< LAInterface > >( 3, std::move( mechPrecond ) ) );
 
-    auto flowPrecond = LAInterface::createPreconditioner( flowSolver()->getLinearSolverParameters() );
+    auto flowPrecond = LAInterface::createPreconditioner( this->flowSolver()->getLinearSolverParameters() );
     precond->setupBlock( 1,
                          { { flow::pressure::key(), { 1, true } } },
                          std::move( flowPrecond ) );
@@ -435,65 +326,13 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::updateState( DomainPartition & dom
                                                               [&]( localIndex const,
                                                                    CellElementSubRegion & subRegion )
     {
-      flowSolver()->updateFluidState( subRegion );
-      if( m_isThermal )
+      this->flowSolver()->updateFluidState( subRegion );
+      if( this->m_isThermal )
       {
-        flowSolver()->updateSolidInternalEnergyModel( subRegion );
+        this->flowSolver()->updateSolidInternalEnergyModel( subRegion );
       }
     } );
   } );
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::mapSolutionBetweenSolvers( DomainPartition & domain, integer const solverType )
-{
-  GEOS_MARK_FUNCTION;
-
-  /// After the flow solver
-  if( solverType == static_cast< integer >( Base::SolverType::Flow ) )
-  {
-    // save pressure and temperature at the end of this iteration
-    flowSolver()->saveIterationState( domain );
-
-    this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                                MeshLevel & mesh,
-                                                                                arrayView1d< string const > const & regionNames )
-    {
-
-      mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                            auto & subRegion )
-      {
-        // update the bulk density
-        // TODO: ideally, we would not recompute the bulk density, but a more general "rhs" containing the body force and the
-        // pressure/temperature terms
-        updateBulkDensity( subRegion );
-
-      } );
-    } );
-  }
-
-  /// After the solid mechanics solver
-  if( solverType == static_cast< integer >( Base::SolverType::SolidMechanics ) )
-  {
-    // compute the average of the mean total stress increment over quadrature points
-    averageMeanTotalStressIncrement( domain );
-
-    this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                                MeshLevel & mesh,
-                                                                                arrayView1d< string const > const & regionNames )
-    {
-
-      mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                            auto & subRegion )
-      {
-        // update the porosity after a change in displacement (after mechanics solve)
-        // or a change in pressure/temperature (after a flow solve)
-        flowSolver()->updatePorosityAndPermeability( subRegion );
-      } );
-    } );
-  }
-  // call base method (needed to perform nonlinear acceleration)
-  Base::mapSolutionBetweenSolvers( domain, solverType );
 }
 
 template< typename FLOW_SOLVER >
@@ -513,48 +352,6 @@ void SinglePhasePoromechanics< FLOW_SOLVER >::updateBulkDensity( ElementSubRegio
     createAndLaunch< parallelDevicePolicy<> >( fluid,
                                                solid,
                                                subRegion );
-}
-
-template< typename FLOW_SOLVER >
-void SinglePhasePoromechanics< FLOW_SOLVER >::averageMeanTotalStressIncrement( DomainPartition & domain )
-{
-  this->template forDiscretizationOnMeshTargets( domain.getMeshBodies(), [&]( string const &,
-                                                                              MeshLevel & mesh,
-                                                                              arrayView1d< string const > const & regionNames )
-  {
-    mesh.getElemManager().forElementSubRegions< CellElementSubRegion >( regionNames, [&]( localIndex const,
-                                                                                          auto & subRegion )
-    {
-      // get the solid model (to access stress increment)
-      string const solidName = subRegion.template getReference< string >( viewKeyStruct::porousMaterialNamesString() );
-      CoupledSolidBase & solid = this->template getConstitutiveModel< CoupledSolidBase >( subRegion, solidName );
-
-      arrayView2d< real64 const > const meanTotalStressIncrement_k = solid.getMeanTotalStressIncrement_k();
-      arrayView1d< real64 > const averageMeanTotalStressIncrement_k = solid.getAverageMeanTotalStressIncrement_k();
-
-      finiteElement::FiniteElementBase & subRegionFE =
-        subRegion.template getReference< finiteElement::FiniteElementBase >( solidMechanicsSolver()->getDiscretizationName() );
-
-      // determine the finite element type
-      finiteElement::FiniteElementDispatchHandler< BASE_FE_TYPES >::
-      dispatch3D( subRegionFE, [&] ( auto const finiteElement )
-      {
-        using FE_TYPE = decltype( finiteElement );
-
-        // call the factory and launch the kernel
-        AverageOverQuadraturePoints1DKernelFactory::
-          createAndLaunch< CellElementSubRegion,
-                           FE_TYPE,
-                           parallelDevicePolicy<> >( mesh.getNodeManager(),
-                                                     mesh.getEdgeManager(),
-                                                     mesh.getFaceManager(),
-                                                     subRegion,
-                                                     finiteElement,
-                                                     meanTotalStressIncrement_k,
-                                                     averageMeanTotalStressIncrement_k );
-      } );
-    } );
-  } );
 }
 
 template class SinglePhasePoromechanics< SinglePhaseBase >;
